Read Discription
Thanks to CPE 353 I now prefer the formatting shown below, so all the if, for, and while loops have been reformatted. More importantly, the command handling has been updated to have the main ESP relay commands for the Motor ESP that it receives. In theory the code for the Motor ESP is done, now the only issue is the Main ESP.

if (condition)
{
   code;
}
else
{
   code;
}